+++ packages/slice-layout/__tests__/s1andc.web.js
@@ -1,5 +1,0 @@
-import { mount } from "enzyme";
-import "./serializers.web";
-import shared from "./s1andc.base";
-
-export default () => shared(mount);