+++ packages/slice-layout/__tests__/sd1and4.web.js
@@ -1,5 +1,0 @@
-import { mount } from "enzyme";
-import "./serializers.web";
-import shared from "./sd1and4.base";
-
-export default () => shared(mount);