+++ packages/slice-layout/__tests__/vertical-list.web.js
@@ -1,5 +1,0 @@
-import { mount } from "enzyme";
-import "./serializers.web";
-import shared from "./vertical-list.base";
-
-export default () => shared(mount);