+++ packages/slice-layout/__tests__/web/clac-with-style.web.test.js
@@ -1,3 +1,0 @@
-import shared from "../clac-with-style.web";
-
-shared();