+++ packages/slice-layout/__tests__/web/l1and1.web.test.js
@@ -1,3 +1,0 @@
-import shared from "../l1and1.web";
-
-shared();