+++ packages/slice-layout/__tests__/web/s1andc.web.test.js
@@ -1,3 +1,0 @@
-import shared from "../s1andc.web";
-
-shared();