+++ packages/slice-layout/__tests__/web/sd2np2-with-style.web.test.js
@@ -1,3 +1,0 @@
-import shared from "../sd2np2-with-style.web";
-
-shared();